scripts/requirements.txt
torch
torchvision
keras
tensorflow
cupy-cuda12x 
numpy
scikit-learn
h5py
matplotlib
seaborn